+++ app/Http/Controllers/InstallationsController.php
@@ -26,6 +26,8 @@ class InstallationsController extends Controller
     /** @var \App\Basket\Synchronisation\InstallationSynchronisationService  */
     private $installationSynchronisationService;
 
+    protected $installationGateway;
+
     /**
      * @author WN
      */
@@ -33,6 +35,9 @@ class InstallationsController extends Controller
     {
         $this->installationSynchronisationService = \App::make(
             'App\Basket\Synchronisation\InstallationSynchronisationService'
+        );
+        $this->installationGateway = \App::make(
+            'PayBreak\Sdk\Gateways\InstallationGateway'
         );
     }
 
@@ -103,6 +108,27 @@ class InstallationsController extends Controller
             'custom_logo_url' => 'url|max:255',
             'location_instruction' => 'max:255',
         ]);
+        $old = new Installation();
+        $old = $old->findOrFail($id);
+
+        if($old->ext_notification_url !== $request->ext_notification_url ||
+            $old->ext_return_url !== $request->ext_return_url) {
+
+            try {
+                $this->installationGateway
+                    ->patchInstallation(
+                        $this->fetchInstallation($id)->ext_id,
+                        [
+                            'return_url' => $request->ext_return_url,
+                            'notification_url' => $request->ext_notification_url
+                        ],
+                        $this->fetchInstallation($id)->merchant->token
+                    );
+            } catch (\Exception $e) {
+                dd($e);
+                return RedirectException::make('/installations/' . $id . '/edit')->setError($e->getMessage());
+            }
+        }
 
         return $this->updateModel((new Installation()), $id, 'installation', '/installations', $request);
     }