+++ resources/views/locations/create.blade.php
@@ -62,6 +62,10 @@
                         regexp: {
                             regexp: '^[A-Za-z0-9\-]+$',
                             message: 'The location reference can only contain letters, numbers and underscores'
+                        },
+                        stringLength: {
+                            max: 242,
+                            message: 'The location reference must not be greater than 242 characters'
                         }
                     }
                 },
@@ -69,6 +73,10 @@
                     validators: {
                         notEmpty: {
                             message: 'The name cannot be empty'
+                        },
+                        stringLength: {
+                            max: 255,
+                            message: 'The name must not be greater than 255 characters'
                         }
                     }
                 },
@@ -77,13 +85,21 @@
                         notEmpty: {
                             message: 'The email address cannot be empty'
                         },
-                        emailAddress: {}
+                        emailAddress: {},
+                        stringLength: {
+                            max: 255,
+                            message: 'The email must not be greater than 255 characters'
+                        }
                     }
                 },
                 address: {
                     validators: {
                         notEmpty: {
                             message: 'The address cannot be empty'
+                        },
+                        stringLength: {
+                            max: 255,
+                            message: 'The address must not be greater than 255 characters'
                         }
                     }
                 }